lambda_function/requirements.txt
google-cloud-bigquery==3.13.0
google-auth==2.23.4
google-api-python-client==2.106.0
requests==2.31.0
python-dateutil==2.8.2